requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.3
chardet==4.0.0
colorama==0.4.4
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.2
Django==3.1.3
django-admin-sortable2==0.7.7
django-cors-headers==3.5.0
django-email-extras==0.3.4
django-extensions==3.0.8
django-filter==2.4.0
django-health-check==3.14.3
django-jsonfield-backport==1.0.3
django-mysql==3.11.1
django-rest-knox==4.1.0
django-shell-plus==1.1.7
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
et-xmlfile==1.0.1
filelock==3.0.12
future==0.15.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
lazy-object-proxy==1.4.3
matplotlib-inline==0.1.2
mysqlclient==2.0.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.9.0
python-gnupg==0.4.7
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
wrapt==1.12.1
